just 2 more extra chapter and done

+++ classes/Ijdb/Controllers/Joke.php
@@ -3,16 +3,22 @@
 namespace Ijdb\Controllers;
 
 use \Ninja\DatabaseTable;
+use \Ninja\Authentication;
 
 class Joke
 {
     private $authorsTable;
     private $jokesTable;
+    private $authentication;
 
-    public function __construct(DatabaseTable $jokesTable, DatabaseTable $authorsTable)
-    {
+    public function __construct(
+        DatabaseTable $jokesTable,
+        DatabaseTable $authorsTable,
+        Authentication $authentication
+    ) {
         $this->jokesTable = $jokesTable;
         $this->authorsTable = $authorsTable;
+        $this->authentication = $authentication;
     }
 
     public function list()
@@ -29,7 +35,8 @@ class Joke
                 'joketext' => $joke['joketext'],
                 'jokedate' => $joke['jokedate'],
                 'name' => $author['name'],
-                'email' => $author['email']
+                'email' => $author['email'],
+                'authorid' => $author['id']
             ];
         }
 
@@ -37,18 +44,15 @@ class Joke
 
         $totalJokes = $this->jokesTable->total();
 
-		ob_start();
-
-		include  __DIR__ . '/../../templates/';
-
-		$output = ob_get_clean();
+        $author = $this->authentication->getUser();
 
         return [
             'template' => 'jokes.html.php',
             'title' => $title,
             'variables' => [
                 'totalJokes' => $totalJokes,
-                'jokes' => $jokes
+                'jokes' => $jokes,
+                'userId' => $author['id'] ?? null
             ]
         ];
     }
@@ -62,6 +66,15 @@ class Joke
 
     public function delete()
     {
+
+        $author = $this->authentication->getUser();
+
+        $joke = $this->jokesTable->findById($_POST['id']);
+
+        if ($joke['authorid'] != $author['id']) {
+            return;
+        }
+
         $this->jokesTable->delete($_POST['id']);
 
         header('location: /joke/list');
@@ -69,8 +82,18 @@ class Joke
 
     public function saveEdit()
     {
+        $author = $this->authentication->getUser();
+
+        if (isset($_GET['id'])) {
+            $joke = $this->jokesTable->findById($_GET['id']);
+
+            if ($joke['authorid'] != $author['id']) {
+                return;
+            }
+        }
+
         $joke = $_POST['joke'];
-        $joke['authorid'] = 1;
+        $joke['authorid'] = $author['id'];
         $joke['jokedate'] = new \DateTime();
 
         $this->jokesTable->save($joke);
@@ -80,6 +103,8 @@ class Joke
 
     public function edit()
     {
+        $author = $this->authentication->getUser();
+
         if (isset($_GET['id'])) {
             $joke = $this->jokesTable->findById($_GET['id']);
         }
@@ -90,7 +115,8 @@ class Joke
             'template' => 'editjoke.html.php',
             'title' => $title,
             'variables' => [
-                'joke' => $joke ?? null
+                'joke' => $joke ?? null,
+                'userId' => $author['id'] ?? null
             ]
         ];
     }
